refactor(handlers): handlers package removed
